+++ src/com/intellij/idea/plugin/hybris/project/configurators/ConfiguratorFactory.kt
@@ -55,7 +55,7 @@ class ConfiguratorFactory {
     fun getLoadedConfigurator(): LoadedConfigurator = ApplicationManager.getApplication().getService(LoadedConfigurator::class.java)
 
     fun getAntConfigurator(): AntConfigurator? =
-        if (PluginCommon.isPluginActive(PluginCommon.ANT_SUPPORT_PLUGIN_ID)) ApplicationManager.getApplication().getService(AntConfigurator::class.java)
+        if (PluginCommon.isPluginActive(PluginCommon.PLUGIN_ANT_SUPPORT)) ApplicationManager.getApplication().getService(AntConfigurator::class.java)
         else null
 
     fun getDataSourcesConfigurator(): DataSourcesConfigurator? = ApplicationManager.getApplication().getService(DataSourcesConfigurator::class.java)